% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_rogue_allele}
\alias{add_rogue_allele}
\title{Add a rogue allele column}
\usage{
add_rogue_allele(x, ms = NULL, name = "rare")
}
\arguments{
\item{x}{a \code{\link{freqt}} object}

\item{ms}{a marker or marker vector. If left undefined, then all markers are
processed.}

\item{name}{the name to give to the rogue allele}
}
\description{
Adds an extra allele to every marker with frequency equal to 1 minus the sum
of the frequencies of the alleles in that marker. Note: if the frequencies
sum up to greater than one, then the rogue allele column will have negative
frequencies.
}
\examples{
add_rogue_allele(ft_popstr_europe)
}
