% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{alleles.freqt}
\alias{alleles.freqt}
\title{Get alleles of marker}
\usage{
alleles.freqt(x, marker)
}
\arguments{
\item{x}{a \code{freqt} object}

\item{marker}{a marker name or index}
}
\value{
a list of allele names
}
\description{
Returns all alleles found in the given \code{\link{freqt}} object for which
the frequency is greater than 0.
}
