% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_nist}
\alias{data_nist}
\alias{ft_nist_caucasian}
\alias{ft_nist_asian}
\alias{ft_nist_african_american}
\alias{ft_nist_hispanic}
\alias{ft_nist_all}
\title{US Allele frequencies}
\format{An object of class \code{freqt} of length 4.}
\source{
\url{https://strbase.nist.gov/NISTpop.htm}

Hill, C. R., Duewer, D. L., Kline, M. C., Coble, M. D., & Butler, J.
  M. (2013). US population data for 29 autosomal STR loci. Forensic Science
  International: Genetics, 7(3), e82-e83.
}
\usage{
ft_nist_caucasian

ft_nist_asian

ft_nist_african_american

ft_nist_hispanic

ft_nist_all
}
\description{
Data extracted from the NIST 1036 Revised U.S. Population Dataset (July
2017). The following collections are available:
}
\details{
\describe{

\item{ft_nist_all}{All data (average over the 1036 individuals)}

\item{ft_nist_asian}{Asian}

\item{ft_nist_african_american}{African American}

\item{ft_nist_caucasian}{Caucasian}

\item{ft_nist_hispanic}{Hispanic}

}
}
\keyword{datasets}
