% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_popstr}
\alias{data_popstr}
\alias{ft_popstr_europe}
\alias{ft_popstr_nw_spain}
\alias{ft_popstr_israel_carmel_druze}
\alias{ft_popstr_cambodia_cambodian}
\alias{ft_popstr_china_dai}
\alias{ft_popstr_china_daur}
\alias{ft_popstr_china_han}
\alias{ft_popstr_japan_japanese}
\alias{ft_popstr_siberia_yakut}
\alias{ft_popstr_central_african_republic_biaka_pygmies}
\alias{ft_popstr_d_r_congo_mbuti_pygmies}
\alias{ft_popstr_kenya_bantu_ne}
\alias{ft_popstr_namibia_san}
\alias{ft_popstr_nigeria_yoruba}
\alias{ft_popstr_senegal_mandenka}
\alias{ft_popstr_somalia}
\alias{ft_popstr_south_africa_bantu}
\alias{ft_popstr_brazil_karitiana}
\alias{ft_popstr_brazil_surui}
\alias{ft_popstr_colombia_colombian}
\alias{ft_popstr_dominican_republic}
\alias{ft_popstr_mexico_maya}
\alias{ft_popstr_mexico_pima}
\alias{ft_popstr_israel_central_palestinian}
\alias{ft_popstr_israel_negev_bedouin}
\alias{ft_popstr_bougainville_nan_melanesian}
\alias{ft_popstr_new_guinea_papuan}
\alias{ft_popstr_china_uygur}
\alias{ft_popstr_pakistan_balochi}
\alias{ft_popstr_pakistan_burusho}
\alias{ft_popstr_pakistan_hazara}
\alias{ft_popstr_pakistan_kalash}
\alias{ft_popstr_pakistan_makrani}
\alias{ft_popstr_pakistan_pathan}
\alias{ft_popstr_pakistan_sindhi}
\alias{ft_popstr_algeria_mzab_mozabite}
\alias{ft_popstr_china_hezhen}
\alias{ft_popstr_china_lahu}
\alias{ft_popstr_china_miaozu}
\alias{ft_popstr_china_mongola}
\alias{ft_popstr_china_naxi}
\alias{ft_popstr_china_oroqen}
\alias{ft_popstr_china_she}
\alias{ft_popstr_china_tu}
\alias{ft_popstr_china_tujia}
\alias{ft_popstr_china_xibo}
\alias{ft_popstr_china_yizu}
\title{pop.STR allele frequencies}
\format{An object of class \code{freqt} of length 6.}
\source{
\url{http://spsmart.cesga.es/popstr.php}

Amigo J, Phillips C, Salas T, Fernández Formoso L, Carracedo A, Lareu
  M. pop.STR - An online population frequency browser for established and new
  forensic STRs. Forensic Sci. Int. Gene. Suppl. 2009
  (10.1016/j.fsigss.2009.08.178)
}
\usage{
ft_popstr_europe

ft_popstr_nw_spain

ft_popstr_israel_carmel_druze

ft_popstr_cambodia_cambodian

ft_popstr_china_dai

ft_popstr_china_daur

ft_popstr_china_han

ft_popstr_japan_japanese

ft_popstr_siberia_yakut

ft_popstr_central_african_republic_biaka_pygmies

ft_popstr_d_r_congo_mbuti_pygmies

ft_popstr_kenya_bantu_ne

ft_popstr_namibia_san

ft_popstr_nigeria_yoruba

ft_popstr_senegal_mandenka

ft_popstr_somalia

ft_popstr_south_africa_bantu

ft_popstr_brazil_karitiana

ft_popstr_brazil_surui

ft_popstr_colombia_colombian

ft_popstr_dominican_republic

ft_popstr_mexico_maya

ft_popstr_mexico_pima

ft_popstr_israel_central_palestinian

ft_popstr_israel_negev_bedouin

ft_popstr_bougainville_nan_melanesian

ft_popstr_new_guinea_papuan

ft_popstr_china_uygur

ft_popstr_pakistan_balochi

ft_popstr_pakistan_burusho

ft_popstr_pakistan_hazara

ft_popstr_pakistan_kalash

ft_popstr_pakistan_makrani

ft_popstr_pakistan_pathan

ft_popstr_pakistan_sindhi

ft_popstr_algeria_mzab_mozabite

ft_popstr_china_hezhen

ft_popstr_china_lahu

ft_popstr_china_miaozu

ft_popstr_china_mongola

ft_popstr_china_naxi

ft_popstr_china_oroqen

ft_popstr_china_she

ft_popstr_china_tu

ft_popstr_china_tujia

ft_popstr_china_xibo

ft_popstr_china_yizu
}
\description{
Data was generated by selecting all possible markers and the corresponding
population (meta-)group in the pop.STR tool.
}
\keyword{datasets}
