% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_strider}
\alias{data_strider}
\alias{ft_strider_austria}
\alias{ft_strider_belgium}
\alias{ft_strider_bosnia_herzegowina}
\alias{ft_strider_czech_republic}
\alias{ft_strider_denmark}
\alias{ft_strider_finland}
\alias{ft_strider_france}
\alias{ft_strider_germany}
\alias{ft_strider_greece}
\alias{ft_strider_hungary}
\alias{ft_strider_ireland}
\alias{ft_strider_montenegro}
\alias{ft_strider_norway}
\alias{ft_strider_poland}
\alias{ft_strider_slovakia}
\alias{ft_strider_slovenia}
\alias{ft_strider_spain}
\alias{ft_strider_sweden}
\alias{ft_strider_switzerland}
\title{STRidER 2.0}
\format{An object of class \code{freqt} of length 5.}
\source{
\url{https://strider.online/frequencies}

Bodner M., Bastisch, I., Butler, J.M., Fimmers, R., Gill, P., Gusmão,
  L., Morling, N., Phillips, C., Prinz, M., Schneider, P.M., Parson, W.
  (2016), 'Recommendations of the DNA Commission of the International Society
  for Forensic Genetics (ISFG) on quality control of autosomal Short Tandem
  Repeat allele frequency databasing (STRidER).', Forensic Sci. Int. Genet.
  24, 97-102
}
\usage{
ft_strider_austria

ft_strider_belgium

ft_strider_bosnia_herzegowina

ft_strider_czech_republic

ft_strider_denmark

ft_strider_finland

ft_strider_france

ft_strider_germany

ft_strider_greece

ft_strider_hungary

ft_strider_ireland

ft_strider_montenegro

ft_strider_norway

ft_strider_poland

ft_strider_slovakia

ft_strider_slovenia

ft_strider_spain

ft_strider_sweden

ft_strider_switzerland
}
\description{
As per the STRidER website, be advised that in these tables, 1 represents all
rare alleles shorter than the accepted allele categories. The value 99
represents all rare alleles longer than the accepted categories.
}
\keyword{datasets}
