% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{exporting_data}
\alias{exporting_data}
\alias{write_familias}
\title{Write a frequency table}
\usage{
write_familias(x, filepath)
}
\arguments{
\item{x}{a \code{\link{freqt}} object}
}
\description{

}
\details{

}
