% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{fafreqs_widget}
\alias{fafreqs_widget}
\title{Frequency database loader}
\usage{
fafreqs_widget(input, output, session, id)
}
\arguments{
\item{input}{the shiny input object}

\item{output}{the shiny output object}

\item{session}{the shiny session object}

\item{id}{the id given to the input}
}
\description{
Use with the \code{\link{callModule}} function.
}
