% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{fafreqs_widget_input}
\alias{fafreqs_widget_input}
\title{Frequency database loader input}
\usage{
fafreqs_widget_input(
  id,
  allow_fafreqs_markersets = TRUE,
  allow_import_familias = TRUE,
  allow_import_csv = TRUE,
  allow_marker_filtering = TRUE,
  allow_scaling = TRUE,
  allow_rare_allele = TRUE
)
}
\arguments{
\item{id}{a unique identifier per shiny"s documentation}

\item{allow_fafreqs_markersets}{allow the user to select one of the freqyency
tables included in the \code{fafreqs} package}

\item{allow_import_familias}{allow the user to load a familias frequency file}

\item{allow_import_csv}{allow the user to load a custom frequency table}

\item{allow_marker_filtering}{allow the user to select which markers are to
be used out of those present in the selected (or uploaded) frequency
database}

\item{allow_scaling}{allow the user to scale the frequencies so that they sum up to 1}

\item{allow_rare_allele}{allow the user to add an allele so that the frequencies sum up to 1}
}
\description{
Frequency database loader input
}
