% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ferbous2010}
\alias{ferbous2010}
\alias{ft_ferdous2010_bangladesh_chakma}
\alias{ft_ferdous2010_bangladesh_tripura}
\title{Ferdous 2010: Bangladesh}
\format{An object of class \code{freqt} of length 6.}
\source{
Ferdous, A., Ali, M. E., Alam, S., Hasan, M., Hossain, T., &
  Akhteruzzaman, S. (2010). Allele Frequencies of 10 Autosomal STR Loci from
  Chakma and Tripura Tribal Populations in Bangladesh. Molecular biology
  international, 2010.
}
\usage{
ft_ferdous2010_bangladesh_chakma

ft_ferdous2010_bangladesh_tripura
}
\description{
Allele frequencies of ten autosomal STR loci, D3S1358, vWA, D16S539, D2S1338,
D8S1179, D21S11, D18S51, D19S433, TH01, and FGA were investigated in Chakma
and Tripura tribal populations of Bangladesh. In both the populations, all
loci were in Hardy-Weinberg equilibrium except for FGA locus in Chakma and
D21S11 in Tripura. All the loci were highly polymorphic in Chakma population
with an observed heterozygosity  of 0.7 and moderately polymorphic in Tripura
population . However, both the population showed least polymorphism at TH01
locus . A comparison between Chakma and Tripura population data revealed
statistically significant differences in allele frequency distribution for
most of the loci. A similar comparison with the mainstream Bengali population
using previously published data from this lab also showed significant
difference in allele frequency with these two tribal populations.
}
\keyword{datasets}
