% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_markers}
\alias{filter_markers}
\title{Filter markers in a \code{\link{freqt}} object}
\usage{
filter_markers(x, markers)
}
\arguments{
\item{x}{a \code{\link{freqt}} object}

\item{markers}{a list of marker names}
}
\value{
a \code{\link{freqt}} object which only contains those markers in
  \code{markers}
}
\description{
Filter markers in a \code{\link{freqt}} object
}
\examples{
# pop.STR european frequencies for only the marker TPOX
filter_markers(ft_popstr_europe, c('TPOX'))

}
\seealso{
\code{\link{markers}}
}
