% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqt.R
\name{freqt}
\alias{freqt}
\title{Create a \code{freqt} object}
\usage{
freqt(x, name = "")
}
\arguments{
\item{x}{a \code{\link{data.frame}} where cell \code{(i, j)} contains the
frequency of allele \code{j} of marker \code{i}. The row names of \code{x}
should be the marker names.}

\item{name}{(optional, default = "") the name of the frequency database}
}
\value{
a \code{freqt} objet
}
\description{
Create a \code{freqt} object
}
