% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{frequencies}
\alias{frequencies}
\title{Get allele frequencies}
\usage{
frequencies(x, ms, als = NULL)
}
\arguments{
\item{x}{a \code{\link{freqt}} object}

\item{ms}{a marker or marker vector}

\item{als}{an allele or allele vector. If none is provided all relevant
alleles will be returned for the selected markers.}
}
\value{
depending on the number of alleles provided a single number or a data
  frame is returned with the request allele frequencies
}
\description{
Get allele frequencies
}
\examples{
# a single marker and allele returns a number
frequencies(ft_popstr_europe, "FGA", 23)

# multiple alleles result in a data.frame
frequencies(ft_popstr_europe, "FGA", c(6, 23, 7, 23.2))

# multiple markers and alleles also result in a data.frame
frequencies(ft_popstr_europe, c("CSF1PO", "FGA"), c(6, 23, 7, 23.2))

# if alleles are not specified then every non-zero frequency is returned
frequencies(ft_popstr_europe, "FGA")

# similarly, if multiple markers are provided any allele
# with a non-zero frequency in any of the markers is returned
frequencies(ft_popstr_europe, c("CSF1PO", "FGA"))
}
