% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ft_cho2003_south_korea}
\alias{ft_cho2003_south_korea}
\title{Cho 2003: South Korea}
\format{An object of class \code{freqt} of length 6.}
\source{
Cho, N. S., Hwang, J. H., Lee, Y. A., & Park, I. H. (2003).
  Population genetics of nine STR loci: TH01, TPOX, CSF1PO, vWA, FESFPS,
  F13A01, D13S317, D7S820 and D16S539 in a Korean population. Forensic
  science international, 137(1), 97-99.
}
\usage{
ft_cho2003_south_korea
}
\description{
Allele frequencies for nine STR loci namely, TH01, TPOX, CSF1PO, vWA, FESFPS,
F13A01, D13S317, D7S820 and D16S539 were obtained from a sample of 437
unrelated individuals living in Chungcheong-do, South Korea.
}
\keyword{datasets}
