% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ft_zeyad2020_uae_ystrs}
\alias{ft_zeyad2020_uae_ystrs}
\title{Zeyad 2020: United Arab Emirates Y-STRs}
\format{An object of class \code{freqt} of length 6.}
\source{
Zeyad, T., Adam, A., Alghafri, R., & Iratni, R. (2020). Study of 27
  Y-STR markers in United Arab Emirates population. Forensic Science
  International: Reports, 100057.
}
\usage{
ft_zeyad2020_uae_ystrs
}
\description{
The recently introduced 6-dyes Yfiler® Plus multiplex kit which includes 27
Y-STR loci has been used to study 436 UAE Arab male individuals. In addition
to the 17 Y-STRs traditionally used in Yfiler® multiplex kit, Yfiler® Plus
has ten additional Y-STRs. It includes seven rapidly mutating Y-STRs (RM
Y-STRs) namely DYS560, DYS576, DYS627, DYS518, DYS449 and DYF387S1ab, and
three moderately mutating Y-STRs namely DYS481, DYS533 and DYS460. The RM
Y-STRs were already proven to be powerful for discriminating closely related
and unrelated males’ individuals. As a result, a total number of 419
haplotypes were observed out of 436 samples analysed of UAE Arab males’
population, whereas 15 haplotypes were shared between two individuals and one
haplotypes were shared between three individuals. The discrimination capacity
(DC) for the 27 loci was found to be 0.9618 whereas haplotype diversity (HD)
was found to be 0.99982. AMOVA results showed that UAE population clustered
along with Saudi Arabia East [Arab] population which indicates that they have
a closer genetic relation. While it was standing away compared to other
relevant populations. In conclusion, Yfiler® Plus multiplex kit showed higher
power of discrimination than all previous studies conducted on the same
population. This study considered the first study conducted using 27 Y-STR
loci in this population. Geographic distribution of the populations analysed
in this study are well reflected in the genetic distance results.
}
\keyword{datasets}
