% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kits.R
\name{get_marker_kit}
\alias{get_marker_kit}
\title{Standard Amplification Kits}
\usage{
get_marker_kit(abbrev = NULL, name = NULL, just_names = FALSE)
}
\arguments{
\item{abbrev}{abbreviation code for the kit}

\item{name}{name of the kit}

\item{just_names}{wether to return the abbreviation of the matching kits in a
named list or just the list of markers amplified by that kit.}
}
\value{
a (list of) kit which consists in a list of markers (if
  \code{just_names = FALSE}) or the abbreviation of the kit. If multiple kits
  match or no filtering parameters were provided a list of kits is returned.
}
\description{
This function returns information about the most comonly used amplification
kits. The \code{abbrev} and \code{name} params allow the user to search for
kits in this package. The \code{just_names} param controls the output of the
function. If no parameters are provided, it returns a list of all the kits
that this package includes.
}
\examples{
# get a list of the names of all the available kits
get_marker_kit(just_names = TRUE)

# get the markers included in the "Qiagen Investigator HDplex" kit
get_marker_kit(name = "Qiagen Investigator HDplex")

# get the markers included in the "core23" (abbrev) kit
# this is useful for when you know the abbreviation
# of a kit which is shorter to type
get_marker_kit(abbrev = "core23")
}
