% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{guess_chromosome_numbers}
\alias{guess_chromosome_numbers}
\title{Guess chromosome numbers}
\usage{
guess_chromosome_numbers(x, as_numeric = FALSE)
}
\arguments{
\item{x}{a \code{\link{freqt}} object}

\item{as_numeric}{whether to convert chromosome names "X" and "Y" to the
number 23}
}
\value{
an updated version of \code{x} with the known chormosome numbers attached
}
\description{
Guess chromosome numbers
}
\examples{
guess_chromosome_numbers(ft_popstr_cambodia_cambodian)
}
