% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{marker_metadata}
\alias{marker_metadata}
\title{Marker metadata}
\format{An object of class \code{data.frame} with 136 rows and 1 columns.}
\source{
https://strider.online/bundles/strbaseclient/downloads/Forensic_STR_Sequence_Structure_Guide_v5.xlsx
}
\usage{
marker_metadata
}
\description{
A \code{data.frame} containing marker metadata (currently only chromosome
number)
}
\keyword{datasets}
