% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_sets.R
\docType{data}
\name{marker_sets}
\alias{marker_sets}
\alias{core23}
\alias{core24}
\alias{illumina_forenseq}
\alias{qiagen_investigator}
\alias{promega_cs7}
\alias{usc_aim}
\alias{nist_mini}
\title{Common Marker Sets}
\format{An object of class \code{character} of length 23.}
\usage{
core23

core24

illumina_forenseq

qiagen_investigator

promega_cs7

usc_aim

nist_mini
}
\description{
Common Marker Sets
}
\keyword{datasets}
