% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{markers}
\alias{markers}
\title{Get markers in a \code{\link{freqt}} object}
\usage{
markers(x)
}
\arguments{
\item{x}{a \code{\link{freqt}} object}
}
\value{
a character vector containing the markers for which frequency data is
  available in \code{x}
}
\description{
Get markers in a \code{\link{freqt}} object
}
\examples{
markers(ft_popstr_europe)

}
\seealso{
\code{\link{filter_markers}}
}
