% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalise}
\alias{normalise}
\title{Normalise a frequency database}
\usage{
normalise(x, ms = NULL)
}
\arguments{
\item{x}{a \code{\link{freqt}} object}

\item{ms}{a marker or a marker vector. If not provided, all markers are
scaled.}
}
\value{
a scaled version of the given frequency database
}
\description{
Scales a frequency database so that allele frequencies sum up to one.
}
\examples{
n <- normalise(ft_popstr_europe)

# this is useful to check, but, in general, you should not
# depend on the internal structure of a freqt object
rowSums(n$TABLE, na.rm = TRUE)
}
