% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqt.R
\name{print.freqt}
\alias{print.freqt}
\title{Printing frequency tables}
\usage{
\method{print}{freqt}(x, ...)
}
\arguments{
\item{x}{a `freqt` object}

\item{...}{further parameters}
}
\description{
Print a `freqt` object
}
\details{
This first calls [as.data.frame.freqt()] and then prints the resulting
`data.frame`. The `data.frame` is returned invisibly.
}
