% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_allelic_ladder}
\alias{read_allelic_ladder}
\title{Read data in allelic ladder format}
\usage{
read_allelic_ladder(filename, name = "", source = NULL, ...)
}
\arguments{
\item{filename}{a filepath}

\item{name}{the name of the dataset}

\item{source}{the source of the dataset}

\item{...}{further parameters to \code{read.csv}}
}
\value{
a \code{\link{freqt}} object
}
\description{
Read data in allelic ladder format
}
