% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_all_na_columns}
\alias{remove_all_na_columns}
\title{Remove columns where every row is NA}
\source{
\url{https://stackoverflow.com/a/2644009/1646268}
}
\usage{
remove_all_na_columns(x)
}
\arguments{
\item{x}{a dataframe}
}
\value{
`x` but with columns where every row was NA removed
}
\description{
Remove columns where every row is NA
}
