% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rownames_to_column}
\alias{rownames_to_column}
\title{Convert rownames to the first column}
\usage{
rownames_to_column(x, name = "X")
}
\arguments{
\item{x}{a \code{\link[base]{data.frame}}}

\item{name}{the column name for the row names}
}
\description{
Convert rownames to the first column
}
