% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-table-file-loader.R
\name{shinyTableFileLoaderInput}
\alias{shinyTableFileLoaderInput}
\title{Table File Loader Input}
\usage{
shinyTableFileLoaderInput(id, label = "Table loader")
}
\arguments{
\item{id}{a stirng that is unique among the widgets used in the interface}

\item{label}{the string shown as a link to open the modal dialog for loading
the file}
}
\description{
UI function for the Table File Loader shiny module.
}
\details{
The Table File Loader shiny module makes it easy for users to load
tabular data from delimited files by allowing them to customise loading
parameters such as column separator, header presence and more. When invoked,
it shows an interactive modal with a preview of the table that is being
loaded that is updated when the user modifies loading settings. The module
returns a dataframe from the data loaded using the user-specified parameters.

This UI function must be included inside the `ui.R` file or `ui()` function of
the *shiny* interface. See examples. Also, the server part of this module,
[shinyTableFileLoader()] must be included in the server section of the
shiny application.
}
\seealso{
[shinyTableFileLoader()] for the server part of this module
}
