% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{standard_marker_names}
\alias{standard_marker_names}
\title{Standarise marker names}
\usage{
standard_marker_names(ms)
}
\arguments{
\item{ms}{a marker name (or a character vector of marker names)}
}
\value{
an updated version of \code{ms} where the names have been standarised
}
\description{
Standarise marker names
}
\examples{
standard_marker_names("vWA") == "VWA"

standard_marker_names(c("Penta D", "vWA")) == c("Penta_D", "VWA")
}
\seealso{
\code{\link{standarise_names}}
}
