% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{standarise_names}
\alias{standarise_names}
\title{Standarise marker names}
\usage{
standarise_names(x)
}
\arguments{
\item{x}{a \code{\link{freqt}} object}
}
\value{
an updated version of \code{x} where the marker names have been standarised
}
\description{
Standarise marker names
}
\examples{
# vWA is changed to VWA
markers(standarise_names(ft_nist_african_american))
}
\seealso{
\code{\link{standard_marker_names}}
}
