% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_pedtools_locusAttributes}
\alias{to_pedtools_locusAttributes}
\title{Convert a \code{\link{freqt}} object to the pedtools \code{locusAttributes}
format.}
\usage{
to_pedtools_locusAttributes(x, scale = TRUE)
}
\arguments{
\item{x}{a \code{freqt} object}

\item{scale}{wether to scale allele frequencies so that they sum up to one
(which is required by pedtools)}
}
\value{
a \code{pedtools}-compatible list of lists following the
  \code{locusAttributes} definition
}
\description{
This function is useful for loading the allele frequency databases found in
this package into pedtools or other pedsuite packages which use the pedtools
package such as forrel.
}
\examples{
library(pedtools)
p = nuclearPed(1)
p = setMarkers(p, locusAttributes = to_pedtools_locusAttributes(ft_nist_african_american))

}
\seealso{
\code{\link{to_pedtools_markers}}

\code{\link[pedtools]{marker_attach}}
}
