% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_pedtools_markers}
\alias{to_pedtools_markers}
\title{Convert a \code{\link{freqt}} object to a list of marker objects}
\usage{
to_pedtools_markers(x, ped, scale = TRUE)
}
\arguments{
\item{x}{a \code{freqt} object}

\item{ped}{a \code{\link[pedtools]{ped}} object}

\item{scale}{wether to scale allele frequencies so that they sum up to one (which is required by pedtools)}
}
\value{
a list of \code{\link[pedtools]{marker}} objects
}
\description{
This function converts a frequency table object into a list of pedtools \code{\link[pedtools]{marker}} objects.
}
\examples{
library(pedtools)
p = nuclearPed(1)
p = setMarkers(p, m = to_pedtools_markers(ft_nist_african_american, p))

}
\seealso{
\code{\link{to_pedtools_locusAttributes}}

\code{\link[pedtools]{marker_attach}}
}
