% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqt.R
\name{as.data.frame.freqt}
\alias{as.data.frame.freqt}
\title{Convert a \code{\link{freqt}} to a \code{\link{data.frame}}}
\usage{
\method{as.data.frame}{freqt}(x, ..., markers = NULL)
}
\arguments{
\item{x}{a \code{\link{freqt}} object}

\item{...}{further parameters}

\item{markers}{a character vector of markers to be included in the resulting
dataframe. By default, all markers available in the dataset will be
included.}
}
\value{
a \code{\link{data.frame}} with one allele per column and one marker
  per row. The \code{row.names} of the result are the names of the markers.
}
\description{
Convert a \code{\link{freqt}} to a \code{\link{data.frame}}
}
