% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_popstr}
\alias{data_popstr}
\alias{ft_popstr_europe}
\alias{ft_popstr_nw_spain}
\alias{ft_popstr_israel_carmel_druze}
\alias{ft_popstr_cambodia_cambodian}
\alias{ft_popstr_china_dai}
\alias{ft_popstr_china_daur}
\alias{ft_popstr_china_han}
\alias{ft_popstr_japan_japanese}
\alias{ft_popstr_siberia_yakut}
\title{pop.STR allele frequencies}
\format{An object of class \code{freqt} of length 4.}
\source{
\url{http://spsmart.cesga.es/popstr.php}
}
\usage{
ft_popstr_europe

ft_popstr_nw_spain

ft_popstr_israel_carmel_druze

ft_popstr_cambodia_cambodian

ft_popstr_china_dai

ft_popstr_china_daur

ft_popstr_china_han

ft_popstr_japan_japanese

ft_popstr_siberia_yakut
}
\description{
Data was generated by selecting all possible markers and the corresponding
population (meta-)group in the pop.STR tool. The following datasets are
available:
}
\details{
\describe{

\item{ft_popstr_europe}{Europe metaregion (average of all other regions)}

\item{ft_popstr_nw_spain}{North-West Spain (Galicia)}

\item{ft_popstr_israel_carmel_druze}{Israel (Carmel) - Druze}

\item{ft_popstr_cambodia_cambodian}{Cambodia - Cambodian}

\item{ft_popstr_china_dai}{China - Dai}

\item{ft_popstr_china_daur}{China - Daur}

\item{ft_popstr_china_han}{China - Han}

\item{ft_popstr_japan_japanese}{Japan - Japanese}

\item{ft_popstr_siberia_yakut}{Siberia - Yakut}

}
}
\keyword{datasets}
