% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_popstr}
\alias{data_popstr}
\alias{ft_popstr_europe}
\alias{ft_popstr_nw_spain}
\alias{ft_popstr_israel_carmel_druze}
\title{pop.STR frequency data}
\format{An object of class \code{freqt} of length 4.}
\source{
\url{http://spsmart.cesga.es/popstr.php}
}
\usage{
ft_popstr_europe

ft_popstr_nw_spain

ft_popstr_israel_carmel_druze
}
\description{
Data was generated by selecting all possible markers and the corresponding
population (meta-)group in the pop.STR tool. The following datasets are
available:
}
\details{
\describe{

\item{ft_popstr_europe}{Europe metaregion (average of all other regions)}

\item{ft_popstr_nw_spain}{North-West Spain (Galicia)}

\item{ft_popstr_israel_carmel_druze}{Israel (Carmel) - Druze}

}
}
\keyword{datasets}
