% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_strider}
\alias{data_strider}
\alias{ft_strider_austria}
\alias{ft_strider_belgium}
\alias{ft_strider_bosnia_herzegowina}
\alias{ft_strider_czech_republic}
\alias{ft_strider_denmark}
\alias{ft_strider_finland}
\alias{ft_strider_france}
\alias{ft_strider_germany}
\alias{ft_strider_greece}
\alias{ft_strider_hungary}
\alias{ft_strider_ireland}
\alias{ft_strider_montenegro}
\alias{ft_strider_norway}
\alias{ft_strider_poland}
\alias{ft_strider_slovakia}
\alias{ft_strider_slovenia}
\alias{ft_strider_spain}
\alias{ft_strider_sweden}
\alias{ft_strider_switzerland}
\title{STRidER 2.0}
\format{An object of class \code{freqt} of length 2.}
\source{
\url{https://strider.online/frequencies}
}
\usage{
ft_strider_austria

ft_strider_belgium

ft_strider_bosnia_herzegowina

ft_strider_czech_republic

ft_strider_denmark

ft_strider_finland

ft_strider_france

ft_strider_germany

ft_strider_greece

ft_strider_hungary

ft_strider_ireland

ft_strider_montenegro

ft_strider_norway

ft_strider_poland

ft_strider_slovakia

ft_strider_slovenia

ft_strider_spain

ft_strider_sweden

ft_strider_switzerland
}
\description{
As per the STRidER website, be advised that in these tables, 1 represents all
rare alleles shorter than the accepted allele categories. The value 99
represents all rare alleles longer than the accepted categories. The
following population groups are available:
}
\details{
\describe{

\item{ft_strider_austria}{Austria}

\item{ft_strider_belgium}{Belgium}

\item{ft_strider_bosnia_herzegowina}{Bosnia and Herzegowina}

\item{ft_strider_czech_republic}{Czech Republic}

\item{ft_strider_denmark}{Denmark}

\item{ft_strider_finland}{Finland}

\item{ft_strider_france}{France}

\item{ft_strider_germany}{Germany}

\item{ft_strider_greece}{Greece}

\item{ft_strider_hungary}{Hungary}

\item{ft_strider_ireland}{Ireland}

\item{ft_strider_montenegro}{Montenegro}

\item{ft_strider_norway}{Norway}

\item{ft_strider_poland}{Poland}

\item{ft_strider_slovakia}{Slovakia}

\item{ft_strider_slovenia}{Slovenia}

\item{ft_strider_spain}{Spain}

\item{ft_strider_sweden}{Sweden}

\item{ft_strider_switzerland}{Switzerland}

}
}
\keyword{datasets}
