% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{exporting_data}
\alias{exporting_data}
\alias{write_csv}
\alias{write_familias}
\title{Write a frequency table}
\usage{
write_csv(x, filepath, included_markers = NULL)

write_familias(x, filepath, included_markers = NULL)
}
\arguments{
\item{x}{a \code{\link{freqt}} object}

\item{filepath}{the path where to write the output file}

\item{included_markers}{markes to include in the exported file (default to
all available markers)}
}
\description{
Writes a frequency table in one of the implemented formats.
}
\details{
\describe{

\item{write_csv}{Writes to a CSV file in the standard one-marker-per-row,
one-allele-per-column format.}

\item{write_familias}{Writes to a familias allele frequency database file.}

}
}
\seealso{
\link{importing_data}
}
