% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{exporting_data}
\alias{exporting_data}
\alias{write_csv}
\alias{write_familias}
\title{Write a frequency table}
\usage{
write_csv(x, filepath, included_markers = NULL)

write_familias(x, filepath, included_markers = NULL)
}
\arguments{
\item{x}{a \code{\link{freqt}} object}

\item{filepath}{the path where to write the output file}

\item{included_markers}{markes to include in the exported file (default to
all available markers)}
}
\description{

}
\details{

}
