% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqt.R
\name{freqt}
\alias{freqt}
\title{Create a `freqt` object}
\usage{
freqt(x, name = "")
}
\arguments{
\item{x}{a `data.frame` where cell (i, j) contains the frequency of allele j
of marker i. The row names of `x` should be the marker names.}

\item{name}{(optional, default = "") the name of the frequency database}
}
\value{
a `freqt` objet
}
\description{
Create a `freqt` object
}
