% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{importing_data}
\alias{importing_data}
\alias{read_popstr}
\alias{read_nist}
\alias{read_familias}
\title{Parse common allele frequency data formats}
\usage{
read_popstr(filename, name = "")

read_nist(filename, name = "")

read_familias(filepath, name)
}
\arguments{
\item{filename}{the path to the relevant CSV file}

\item{name}{the name to give to the dataset}

\item{origin}{(only for read_strider) which population group to choose}
}
\value{
a \code{\link{freqt}} object with the given data
}
\description{
Reads a tabular data file and generates a \code{\link{freqt}} object.
Supports multiple common data sources:
}
\details{

}
