% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{importing_data}
\alias{importing_data}
\alias{read_popstr}
\alias{read_nist}
\alias{read_familias}
\alias{read_strider}
\title{Parse common allele frequency data formats}
\usage{
read_popstr(filename, name = "")

read_nist(filename, name = "")

read_familias(filename, name = "")

read_strider(filename, name = "", origin = "")
}
\arguments{
\item{filename}{the path to the relevant CSV file}

\item{name}{the name to give to the dataset}

\item{origin}{(only for read_strider) which population group to choose}
}
\value{
a \code{\link{freqt}} object with the given data
}
\description{
Reads a tabular data file and generates a \code{\link{freqt}} object.
Supports multiple common data sources:
}
\details{
\describe{

\item{\code{read_popstr}}{supports reading from CSV files downloaded from the
pop.STR project website. No modification of the files is required.}

\item{\code{read_nist}}{Supports reading from CSV files generated from NIST
STRBase data. NIST data comes in MS Excel files with multiple sheets and
additional formating. This function expexts a CSV file with only one of those
sheets.}

\item{\code{read_familias}}{Supports reading from a Familias frequency
database file. The Familias importer does not support reading the number of
individuals as it is not included in Familias data files.}

\item{\code{read_strider}}{Supports extracting frequency data from the
STRidER 2.0 XML database. Requires that the parameter \code{origin} is
supplied and that it matches one of the keys in the STRidER database.}

}
}
