% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqt.R
\name{print.freqt}
\alias{print.freqt}
\title{Printing frequency tables}
\usage{
\method{print}{freqt}(x, ...)
}
\arguments{
\item{x}{a \code{\link{freqt}} object}

\item{...}{further parameters}
}
\description{
Print a \code{\link{freqt}} object
}
\details{
This first calls \code{\link{as.data.frame.freqt}} and then prints the
resulting \code{\link{data.frame}}. The \code{data.frame} is returned
invisibly.
}
\seealso{
\code{\link{as.data.frame.freqt}}
}
